// src/app/api/model/tokenPair.ts
/**
 * orbisapi
 * Polls is a simple API allowing consumers to view polls and vote in them.
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * An access- and refresh token pair.
 */
export interface TokenPair { 
    /**
     * A JWT token to access the API
     */
    access_token?: string;
    /**
     * A JWT token to refresh the access token to get a new token after expiration.
     */
    refresh_token?: string;
}