// src/app/api/model/problem.ts
/**
 * orbisapi
 * Polls is a simple API allowing consumers to view polls and vote in them.
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Problem { 
    /**
     * the title of the occurred problem
     */
    title?: string;
    /**
     * the more detailed information of the problem
     */
    detail?: string;
    /**
     * the type of the problem
     */
    type?: string;
    /**
     * the instance where the problem occurred
     */
    instance?: string;
    /**
     * the returned http status code for older browsers
     */
    status?: number;
}